requirements.txt
fastapi
uvicorn
diffusers
transformers
torch
torchvision
accelerate
safetensors
Pillow
python-multipart
pydantic
